requirements.txt
-i https://pypi.org/simple
aiofiles==22.1.0 ; python_version >= '3.7' and python_version < '4.0'
aiogram==3.0.0b6
aiohttp==3.8.4 ; python_version >= '3.6'
aiosignal==1.3.1 ; python_version >= '3.7'
aiosqlite==0.17.0 ; python_version >= '3.6'
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==22.2.0 ; python_version >= '3.6'
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
executing==1.2.0
frozenlist==1.3.3 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
iso8601==1.1.0 ; python_version < '4.0' and python_full_version >= '3.6.2'
magic-filter==1.0.9 ; python_version >= '3.7' and python_version < '4.0'
multidict==6.0.4 ; python_version >= '3.7'
pydantic==1.10.7 ; python_version >= '3.7'
pypika-tortoise==0.1.6 ; python_version >= '3.7' and python_version < '4.0'
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.2
requests==2.28.2
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tortoise-orm==0.19.3
typing-extensions==4.5.0 ; python_version >= '3.7'
urllib3==1.26.15 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
varname==0.11.0
yarl==1.8.2 ; python_version >= '3.7'
